% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{\%+=\%}
\alias{\%+=\%}
\alias{\%-=\%}
\alias{\%*=\%}
\alias{\%/=\%}
\alias{\%^=\%}
\alias{\%**=\%}
\alias{\%mod=\%}
\title{Basic arithmetic}
\usage{
a \%+=\% b

a \%-=\% b

a \%*=\% b

a \%/=\% b

a \%^=\% b

a \%**=\% b

a \%mod=\% b
}
\arguments{
\item{a}{a}

\item{b}{a}
}
\description{
Basic arithmetic
}
